решение cange46

+++ task46/Program.cs
@@ -1,20 +1,20 @@
 ﻿// Написать программу масштабирования фигуры
 
-int[,] AddPointToArray() //Ввод точек и преобразование в массив
+double[,] AddPointToArray() //Ввод точек и преобразование в массив
 {
     Console.WriteLine("Введите точки фигуры в виде (0,0) (2,0) (2,2) (0,2)");
     string[] text = Console.ReadLine().Replace(")", "").Replace("(", "").Split(" ");
-    int[,] array = new int[text.Length, 2];
+    double[,] array = new double[text.Length, 2];
     for (int i = 0; i < text.Length; i++)
     {
         string[] point = text[i].Split(",");
-        array[i, 0] = Convert.ToInt32(point[0]);
-        array[i, 1] = Convert.ToInt32(point[1]);
+        array[i, 0] = Convert.ToDouble(point[0]);
+        array[i, 1] = Convert.ToDouble(point[1]);
     }
     return array;
 }
 
-void PrintPoints(int[,] matr) // Вывод точек
+void PrintPoints(double[,] matr) // Вывод точек
 {
     for (int i = 0; i < matr.GetLength(0); i++)
     {
@@ -23,7 +23,7 @@ void PrintPoints(int[,] matr) // Вывод точек
     Console.WriteLine();
 }
 
-int[,] ScaleArray(int[,] matr, int k) // Умножение массива на число
+double[,] ScaleArray(double[,] matr, double k) // Умножение массива на число
 {
     for (int i = 0; i < matr.GetLength(0); i++)
     {
@@ -36,8 +36,8 @@ int[,] ScaleArray(int[,] matr, int k) // Умножение массива на 
 }
 
 Console.WriteLine("Программа масштабирования фигуры");
-int[,] arrayPoint = AddPointToArray();
+double[,] arrayPoint = AddPointToArray();
 Console.Write("Введите коэффициэнт масштабирования:");
-int scale = int.Parse(Console.ReadLine());
+double scale = double.Parse(Console.ReadLine());
 arrayPoint = ScaleArray(arrayPoint, scale);
 PrintPoints(arrayPoint);
